Coding fixes and cleanup.

+++ Contents/Library/debugdir_clean.py
@@ -1,8 +1,8 @@
-#  Copyright (c) 2020. Kenneth A. Grady
+#  Copyright (c) 2021. Kenneth A. Grady
 #  See BSD-2-Clause-Patent license in LICENSE.txt
 #  Additional licenses are in the license folder.
 
-""" As the first step in RtoX, delete files the debugdir directory left from
+""" As the second step in RtoX, delete files the debugdir directory left from
 the prior run. (The files are not deleted at the end of a run so they can be
 accessed for debugging purposes.) """
 
@@ -14,17 +14,20 @@ __date__ = "2020-01-09"
 __name__ = "Contents.Library.debugdir_clean"
 
 # From standard libraries
+import logging
 import os
 import sys
 
+log = logging.getLogger(__name__)
+
 
 def cleaner():
-    # TODO Check to see what happens if debugdir is already empty.
     folder = os.path.join(os.path.dirname(os.path.abspath(sys.argv[0])),
                           "debugdir")
     for filename in os.listdir(folder):
         try:
-            os.remove(os.path.join(folder, filename))
+            if filename != "__pycache__":
+                os.remove(os.path.join(folder, filename))
         except OSError:
-            # TODO Insert logging statement.
-            pass
+            log.debug("Problem cleaning out debug_dir directory.")
+            sys.exit()
